refactor: Extract queries to simplify customization
